backend/requirements.txt
python-dotenv
uvicorn
validators
haystack-ai
markdown-it-py
mdit_plain
chromadb
pinecone
boto3
litellm
redis
snowflake-connector-python
docling
fastapi
pydantic
python-multipart
pinecone_haystack
chroma-haystack
nltk
scipy
mistralai


